+++ Baek_joon_java/src/정렬/소트인사이드.java
@@ -1,4 +1,4 @@
-package 자료구조;
+package 정렬;
 
 import java.util.Arrays;
 import java.util.Scanner;
