+++ Baek_joon_java/src/정렬/수정렬2.java
@@ -1,4 +1,4 @@
-package 자료구조;
+package 정렬;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
