+++ Baek_joon_java/src/탐색/ABCDE.java
@@ -1,4 +1,4 @@
-package 자료구조;
+package 탐색;
 
 import java.io.BufferedReader;
 import java.io.IOException;
