+++ Baek_joon_java/src/탐색/트리지름구하기.java
@@ -1,4 +1,4 @@
-package 자료구조;
+package 탐색;
 
 import java.io.IOException;
 import java.util.ArrayList;
